test(index): extract render logic to root index file
