+++ packages/ui/__tests__/Separators/Separator.test.tsx
@@ -1,5 +1,6 @@
 import { Separator, SeparatorProps } from '../../src';
-import { createRender, getSeparator, otherProps } from '.';
+import { createRender, otherProps } from '../';
+import { getSeparator } from '.';
 
 const renderSeparator = createRender<SeparatorProps>(Separator);
