+++ packages/ui/__tests__/Separators/index.tsx
@@ -1,28 +1,4 @@
-import * as React from 'react';
-import { render, screen, within } from '@testing-library/react';
-
-export const otherProps = {
-  className: 'className',
-  id: 'id',
-  style: { display: 'flex' },
-};
-
-export function createRender<Props>(
-  Element: React.ElementType,
-  defaultProps?: Partial<Props>,
-) {
-  return function renderElement(customProps?: Partial<Props>) {
-    const result = render(<Element {...defaultProps} {...customProps} />);
-
-    function rerender(rerenderProps?: Partial<Props>) {
-      return result.rerender(
-        <Element {...defaultProps} {...customProps} {...rerenderProps} />,
-      );
-    }
-
-    return { ...result, rerender };
-  };
-}
+import { screen, within } from '@testing-library/react';
 
 export function getSeparator() {
   return screen.getByRole('separator');
